// carGUI/src/app/models/carWithTypeDetails.ts
import { CarType } from './carType';

export class CarWithCarTypeDetails {
    carNumber: string;
    carType: number;
    region: string;
    mileage: number;

    manufacturer: string;
    carModel: string;
    dailyCost: number;
    dailyPenalty: number;
    ManufacuringYear: Date;
    gearType: string;

   
}